More Actions
git-svn-id: https://xenitsupport.jira.com/svn/XENMTOA/trunk/move2alf-war@766 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ src/main/java/eu/xenit/move2alf/web/controller/JobController.java
@@ -464,7 +464,9 @@ public class JobController {
 		getPipelineAssembler().assemblePipeline(job);
 		
 		ConfiguredAction firstAction = editedJob.getFirstConfiguredAction();
-		getJobService().deleteAction(firstAction.getId());
+		if(firstAction!=null){
+			getJobService().deleteAction(firstAction.getId());
+		}
 
 		mav.setViewName("redirect:/job/dashboard");
 		return mav;
